feat: Refactor routing and component imports for consistency and clarity
- Updated import paths for components to follow a consistent naming convention, enhancing code readability.
- Removed the deprecated routes.tsx file to streamline routing management.
- Added new AuthenticatedActions and UnauthenticatedActions components to improve user authentication handling.
- Enhanced AuthBox to store user login type in local storage for better session management.
- Introduced new styles for authentication components to improve UI consistency.
