+++ app/components/authFooter.tsx
@@ -1,4 +1,4 @@
-import Badge from "./badge";
+import Badge from "./Badge";
 import styles from "../styles/auth.module.css";
