+++ app/entry.client.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { createRoot } from 'react-dom/client';
 import { RouterProvider } from 'react-router-dom';
 import { AuthProvider } from './contexts/AuthContext';
-import router from './routes/routes';
+import router from './routes/Routes';
 import './app.css';
 import './styles/global.css';
