+++ app/styles/AuthStyles.css
@@ -1,0 +1,215 @@
+/* Shared styles for auth components */
+.header-actions {
+  display: flex;
+  align-items: center;
+  gap: 24px;
+}
+
+/* Unauthenticated action styles */
+.trial-banner {
+  padding: 8px;
+  background: #1A1A1A;
+  border-radius: 8px;
+  display: flex;
+  align-items: center;
+}
+
+.trial-text {
+  color: rgba(255, 255, 255, 0.7);
+  font-size: 0.9rem;
+  margin: 0;
+}
+
+.trial-highlight {
+  color: #F9DC00;
+  font-weight: 500;
+}
+
+.trial-upgrade {
+  color: white;
+  font-size: 0.9rem;
+  font-weight: 500;
+  text-decoration: underline;
+  margin-left: 4px;
+}
+
+.trial-upgrade:hover {
+  color: #F9DC00;
+}
+
+.auth-buttons {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+}
+
+.login-button, .signup-button {
+  padding: 8px 16px;
+  border-radius: 8px;
+  font-size: 0.9rem;
+  font-weight: 500;
+  color: white;
+  border: none;
+  cursor: pointer;
+  white-space: nowrap;
+}
+
+.login-button {
+  background: #282828;
+}
+
+.login-button:hover {
+  background: #333333;
+}
+
+.signup-button {
+  background: #1C75BD;
+}
+
+.signup-button:hover {
+  background: #1667A5;
+}
+
+/* Authenticated action styles */
+.authenticated-actions {
+  display: flex;
+  align-items: center;
+  gap: 12px;
+  position: relative;
+}
+
+.premium-profile-container {
+  display: flex;
+  align-items: center;
+  background-color: #111;
+  border-radius: 20px;
+  padding: 3px 10px 3px 15px;
+  gap: 10px;
+  box-shadow: 0 2px 5px rgba(0, 0, 0, 0.2);
+  position: relative;
+}
+
+.premium-label {
+  color: white;
+  font-size: 14px;
+  font-weight: 500;
+}
+
+.user-avatar-container {
+  display: flex;
+  align-items: center;
+  cursor: pointer;
+  gap: 5px;
+}
+
+.profile-image {
+  width: 28px;
+  height: 28px;
+  border-radius: 50%;
+  object-fit: cover;
+}
+
+.profile-image-placeholder {
+  width: 28px;
+  height: 28px;
+  border-radius: 50%;
+  background-color: #333;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  color: white;
+  font-weight: 500;
+  font-size: 16px;
+}
+
+.dropdown-arrow {
+  color: white;
+  font-size: 10px;
+}
+
+/* Profile dropdown styles */
+.profile-dropdown {
+  position: absolute;
+  top: 45px;
+  right: 0;
+  background-color: #121212;
+  border-radius: 8px;
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.4);
+  width: 180px;
+  overflow: hidden;
+  z-index: 200;
+}
+
+.profile-dropdown-header {
+  padding: 12px 16px;
+  background-color: #282828;
+  display: flex;
+  flex-direction: column;
+  gap: 4px;
+}
+
+.profile-name {
+  color: white;
+  font-size: 14px;
+  font-weight: 500;
+}
+
+.profile-email {
+  color: rgba(255, 255, 255, 0.6);
+  font-size: 12px;
+}
+
+.profile-dropdown-divider {
+  height: 1px;
+  background-color: #333;
+  margin: 0;
+}
+
+.profile-dropdown-menu {
+  list-style: none;
+  padding: 0;
+  margin: 0;
+}
+
+.profile-dropdown-item {
+  padding: 15px 20px;
+  color: white;
+  font-size: 14px;
+  cursor: pointer;
+  transition: background-color 0.2s;
+}
+
+.profile-dropdown-item:hover {
+  background-color: #282828;
+}
+
+.profile-dropdown-item.logout {
+  color: #FF0000;
+}
+
+.profile-dropdown-item.logout:hover {
+  background-color: rgba(255, 0, 0, 0.1);
+}
+
+/* Mobile responsive styles */
+@media (max-width: 768px) {
+  .auth-buttons {
+    flex-direction: column;
+    width: 100%;
+  }
+  
+  .login-button, .signup-button {
+    width: 100%;
+  }
+  
+  .authenticated-actions {
+    flex-direction: column;
+    align-items: flex-start;
+    width: 100%;
+  }
+  
+  .premium-profile-container {
+    width: 100%;
+    justify-content: space-between;
+  }
+}
