--- app/styles/dashboardHeader.module.css
@@ -1,83 +1,0 @@
-.header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 16px 24px;
-  background: #080808;
-  border-bottom: 1px solid rgba(255, 255, 255, 0.1);
-}
-
-.logo {
-  font-size: 24px;
-  font-weight: 500;
-  color: #FFFFFF;
-}
-
-.headerRight {
-  display: flex;
-  align-items: center;
-  gap: 24px;
-}
-
-.headerActions {
-  display: flex;
-  gap: 12px;
-}
-
-.actionButton {
-  padding: 8px 16px;
-  border-radius: 6px;
-  font-size: 14px;
-  font-weight: 500;
-  color: #FFFFFF;
-  background: rgba(255, 255, 255, 0.1);
-  text-decoration: none;
-  transition: all 0.2s ease;
-}
-
-.actionButton:hover {
-  background: rgba(255, 255, 255, 0.15);
-}
-
-.notificationBadge {
-  position: relative;
-  cursor: pointer;
-}
-
-.notificationDot {
-  position: absolute;
-  top: -2px;
-  right: -2px;
-  width: 8px;
-  height: 8px;
-  background: #0066FF;
-  border-radius: 50%;
-}
-
-.userProfile {
-  width: 32px;
-  height: 32px;
-  border-radius: 50%;
-  overflow: hidden;
-}
-
-.userProfile img {
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-}
-
-@media screen and (max-width: 768px) {
-  .header {
-    padding: 12px 16px;
-  }
-
-  .headerRight {
-    gap: 16px;
-  }
-
-  .actionButton {
-    padding: 6px 12px;
-    font-size: 12px;
-  }
-}
