+++ app/styles/header.css
@@ -308,7 +308,7 @@
 }
 
 .user-profile {
-  width: 36px;
+  width: 80px;
   height: 36px;
   border-radius: 50%;
   border: none;
@@ -318,8 +318,8 @@
 }
 
 .user-profile img {
-  width: 100%;
-  height: 100%;
+  /* width: 100%;
+  height: 100%; */
   object-fit: cover;
 }
