--- app/styles/headerActions.module.css
@@ -1,83 +1,0 @@
-.headerActions {
-  display: flex;
-  align-items: center;
-  gap: 24px;
-}
-
-.trialBanner {
-  display: flex;
-  align-items: center;
-}
-
-.trialText {
-  font-size: 14px;
-  color: rgba(255, 255, 255, 0.7);
-}
-
-.trialHighlight {
-  color: #FFFFFF;
-  font-weight: 500;
-}
-
-.trialUpgrade {
-  color: #0066FF;
-  text-decoration: none;
-  margin-left: 4px;
-}
-
-.trialUpgrade:hover {
-  text-decoration: underline;
-}
-
-.authButtons {
-  display: flex;
-  gap: 12px;
-}
-
-.loginButton,
-.signupButton {
-  padding: 8px 16px;
-  border-radius: 6px;
-  font-size: 14px;
-  font-weight: 500;
-  cursor: pointer;
-  transition: all 0.2s ease;
-}
-
-.loginButton {
-  background: transparent;
-  border: 1px solid rgba(255, 255, 255, 0.2);
-  color: #FFFFFF;
-}
-
-.loginButton:hover {
-  background: rgba(255, 255, 255, 0.1);
-}
-
-.signupButton {
-  background: #0066FF;
-  border: none;
-  color: #FFFFFF;
-}
-
-.signupButton:hover {
-  background: #0052CC;
-}
-
-@media screen and (max-width: 768px) {
-  .headerActions {
-    flex-direction: column;
-    align-items: flex-start;
-    gap: 16px;
-    padding: 16px;
-  }
-
-  .authButtons {
-    width: 100%;
-  }
-
-  .loginButton,
-  .signupButton {
-    flex: 1;
-  }
-}
