+++ app/styles/layouts/StudioPageLayout.css
@@ -50,6 +50,7 @@
   overflow: hidden;
   background-color: #121212;
   height: 100%;
+  margin-left: 44px;
 }
 
 /* Right panel - Studio view */
