+++ app/styles/studio/ChatInterface.css
@@ -8,6 +8,8 @@
   border-radius: 4px;
   overflow: hidden;
   box-shadow: 0 4px 16px rgba(0, 0, 0, 0.3);
+  /* margin-left: 44px; */
+
 }
 
 /* Chat Header */
