--- app/styles/studio/StudioButtons.css
@@ -1,83 +1,0 @@
-.studio-buttons-container {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 10px 16px;
-  background-color: #121212;
-  border-top: 1px solid rgba(255, 255, 255, 0.1);
-}
-
-.footer-left, 
-.footer-center, 
-.footer-right {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-}
-
-.footer-button {
-  background: transparent;
-  border: none;
-  color: #FFFFFF;
-  padding: 6px 12px;
-  font-size: 14px;
-  cursor: pointer;
-  border-radius: 4px;
-  display: flex;
-  align-items: center;
-  gap: 6px;
-  transition: background-color 0.2s;
-}
-
-.footer-button:hover {
-  background-color: rgba(255, 255, 255, 0.1);
-}
-
-.reset-button {
-  background-color: rgba(255, 255, 255, 0.1);
-}
-
-.nav-button {
-  width: 32px;
-  height: 32px;
-  padding: 0;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  background-color: rgba(255, 255, 255, 0.05);
-}
-
-.scribble-button {
-  background-color: rgba(255, 255, 255, 0.1);
-}
-
-.create-lesson-button {
-  background-color: #4B5563;
-  color: white;
-  border: none;
-  border-radius: 4px;
-  padding: 8px 16px;
-  font-size: 14px;
-  font-weight: 500;
-  cursor: pointer;
-  transition: background-color 0.2s;
-}
-
-.create-lesson-button:hover {
-  background-color: #374151;
-}
-
-/* Responsive layout */
-@media (max-width: 576px) {
-  .studio-buttons-container {
-    flex-direction: column;
-    gap: 12px;
-  }
-  
-  .footer-left,
-  .footer-center,
-  .footer-right {
-    width: 100%;
-    justify-content: center;
-  }
-}
